+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # CHANGELOG (0.9.X)
 
+## 0.9.13 ()
+
+### Backwards incompatible changes from 0.9.12
+ * None
+
+### Bug fixes
+ * [`PULL-288`](https://github.com/thiagoesteves/deployex/pull/288) Report a deployment as complete only when one has finished, not on a hot upgrade or an application restart
+
+### Enhancements
+ * [`PULL-288`](https://github.com/thiagoesteves/deployex/pull/288) Report the versions a hot upgrade moved between in the completion notification
+
 ## 0.9.12 🚀 (2026-08-12)
 
 ### Backwards incompatible changes from 0.9.11